check if deletion of s3 object is needed

+++ src/main/java/com/maxmlv/responserthyme/controllers/apis/PostApiController.java
@@ -53,7 +53,8 @@ public class PostApiController {
                          @RequestParam("redirect") String redirect) throws IOException {
         Post post = postService.findPostById(post_id);
 
-        awsS3Service.deleteObject(post.getMediaFile().getName());
+        if (post.getMediaFile().getName() != null)
+            awsS3Service.deleteObject(post.getMediaFile().getName());
         mediaFileService.deleteByPost(post);
         replyService.deleteAllRepliesByPost(post);
         commentService.deleteAllCommentsByPost(post);
